fastapi-cloudwatch/pyproject.toml
[tool.poetry]
name = "fastapi-cloudwatch"
version = "0.1.0"
description = ""
authors = ["Sachin Magar <smagar77@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
Inject = "^4.3.1"
SQLAlchemy = "^2.0.4"
boto3 = "^1.26.79"
mysql-connector-python = "^8.0.32"
alembic = "^1.9.4"
celery = "^5.2.7"
rabbitmq = "^0.2.0"
pika = "^1.3.1"
redis = "^4.5.1"
flower = "^1.2.0"

python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
